requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnspython==2.8.0
gunicorn==25.0.3
idna==3.11
packaging==26.0
pillow==12.1.0
pymongo==4.16.0
python-dotenv==1.2.1
reportlab==4.4.9
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
